Display links to articles on main page
Current method of retrieving article titles is not robust; we just increment a counter until there isn't a file with that number. This is probably OK for now, since articles shouldn't be that big. However, we are potentially reading a big file just for the title.

Considered caching all articles at startup.
Rejected caching because it would be too complex for now. Pushed onto notion for final product.